requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
greenlet==2.0.2
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
psycopg==3.1.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
